% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\alias{estimate.besiw}
\title{Point and interval estimation of posterior distributions}
\usage{
estimate(object, ...)

\method{estimate}{besiw}(object, prob = 0.95, orthogonal = FALSE, ...)
}
\arguments{
\item{object}{An object from \strong{eigenGSIW}.}

\item{...}{Additional arguments passed to or from other methods.}

\item{prob}{A numeric scalar in the interval (0,1) giving the target probability content of the intervals. The nominal probability content of the intervals is the multiple of 1/nrow(obj) nearest to prob.}

\item{orthogonal}{A logical value indicating whether to ensure the orthogonality of the posterior mean of the eigenvectors.}
}
\value{
An array (or matrix) containing the lower and upper bounds of the credible interval, as well as the posterior mean.
}
\description{
Compute the point estimate (mean) and equal-tailed credible interval to describe posterior distribution.
}
\examples{

\dontrun{

# generate a spiked covariance matrix:
n <- 50
p <- 100
K <- 3
leading <- c(50,20,10)
remaining <- rep(1, p - K)
Sigma0 <- diag(c(leading, remaining), p)

# generate data
set.seed(413)
X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = Sigma0)

# Compute both posterior means and credible intervals
# for eigenvalues and eigenvectors of a covariance matrix:
H <- 4 * diag(rep(1, p))
iter <- 100000
burnin <- floor(iter / 2)
thin <- 100
res <- besiw::eigenGSIW(X = X, k = K, H = H, nmcmc = iter, nburn = burnin, nthin = thin)
est <- besiw::estimate(res)
}

}
\seealso{
\code{eigenGSIW}
}
\author{
Sewon Park
}
