% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.besiw.R
\name{plot.besiw}
\alias{plot.besiw}
\title{Plot diagnostics for posterior samples using \pkg{bayesplot}}
\usage{
\method{plot}{besiw}(
  object,
  par = c("lambda", "gamma"),
  elem = list(),
  type = c("all", "trace", "hist", "area", "acf")
)
}
\arguments{
\item{object}{An object of class \code{"besiw"}, typically returned from \code{eigenGSIW()}.}

\item{par}{Character, either \code{"lambda"} (default) to plot eigenvalues or \code{"gamma"} to plot eigenvector elements.}

\item{elem}{A list of indices to specify which elements to plot.
For \code{par = "lambda"}, supply a list like \code{list(1, 3)} to plot the 1st and 3rd eigenvalues.  
For \code{par = "gamma"}, supply indices as pairs like \code{list(c(1,1), c(2,3))} to indicate matrix positions.

If \code{elem} is not specified, then:
for \code{par = “lambda”}, the top \eqn{k} eigenvalues are plotted;
for \code{par = “gamma”}, the diagonal elements \eqn{(i,i)} for \eqn{i = 1,\ldots,k} of the eigenvector matrix are plotted by default.}

\item{type}{Type of plot to generate. One of \code{"trace"}, \code{"hist"}, \code{"area"}, \code{"acf"}, or \code{"all"} (default).
If \code{"all"}, all four types of diagnostics will be displayed in a 2×2 layout using \pkg{patchwork}.}
}
\value{
A ggplot object (or a patchwork layout of multiple ggplots) displaying the requested diagnostic plot(s).
}
\description{
Provides visual diagnostics of MCMC samples for eigenvalues and eigenvectors, including trace plots, histograms, posterior densities, and autocorrelations.
}
\details{
The function supports plotting the posterior samples of either the top \eqn{k} eigenvalues (\code{par = "lambda"}) or the entries of the corresponding eigenvectors (\code{par = "gamma"}).
Users can select specific elements to visualize, and combine multiple diagnostic plots using \code{type = "all"}.
}
\examples{

\dontrun{
# generate a spiked covariance matrix:
n <- 50
p <- 100
K <- 3
leading <- c(50,20,10)
remaining <- rep(1, p - K)
Sigma0 <- diag(c(leading, remaining), p)

# generate data
set.seed(413)
X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = Sigma0)

# estimate eigenvalues and eigenvectors using a generalized shrinkage inverse-Wishart (gSIW) prior:
H <- 4 * diag(rep(1, p))
iter <- 100000
burnin <- floor(iter / 2)
thin <- 100
res <- besiw::eigenGSIW(X = X, k = K, H = H, nmcmc = iter, nburn = burnin, nthin = thin)

# full diagnostics (trace, histogram, density (area), and autocorrelation) for MCMC samples of the k eigenvalues of the covariance matrix
plot(res, par = 'lambda', type = 'all')

# trace plot for MCMC samples of the first and third eigenvalues
plot(res, par = 'lambda', elem = list(1,3), type = 'trace')

# full diagnostics (trace, histogram, density (area), and autocorrelation) for the diagonal elements of the orthogonal matrix of k eigenvectors
plot(res, par = 'gamma', type = 'all')

# autocorrelation plot for the (1,1) and (2,3) elements of the orthogonal matrix of k eigenvectors
plot(res, par = 'gamma', elem = list(c(1,1), c(2,3)), type = 'acf')
}
}
\author{
Sewon Park
}
