% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_k.R
\name{select_k}
\alias{select_k}
\title{Select the optimal number of spiked eigenvalues via WAIC}
\usage{
select_k(X, kmax = NULL, ncores = 2, ...)
}
\arguments{
\item{X}{An \eqn{n \times p} data matrix.}

\item{kmax}{An integer specifying the maximum number of components to be considered. It is automatically adjusted to be no larger than \code{min(n, p) - 1}.}

\item{ncores}{An integer specifying the number of CPU cores for parallel computation. Default is 2.}

\item{...}{Additional arguments passed to \code{eigenGSIW()}, such as prior settings or sampling options.}
}
\value{
An integer giving the selected number of spiked eigenvalues \eqn{k} that minimizes WAIC.
}
\description{
Selects the number of spiked eigenvalues \eqn{k} in a covariance matrix by computing the Watanabe-Akaike Information Criterion (WAIC) 
under the generalized shrinkage inverse-Wishart (gSIW) prior.
}
\details{
The function evaluates a set of candidate values for \eqn{k} by fitting spiked covariance models 
using the gSIW prior through \code{eigenGSIW()}, and computing their WAIC scores.
The value of \eqn{k} that yields the lowest WAIC is selected as the optimal number of spiked eigenvalues.
}
\examples{

\dontrun{
# generate a spiked covariance matrix:
n <- 10
p <- 20
k <- 3
Sigma0 <-diag(c(100,80,60,rep(1,p-k)))

# generate data
set.seed(413)
X <- MASS::mvrnorm(n, mu=rep(0,p),Sigma = Sigma0)

# Determine the number of spiked eigenvalues k in the covariance matrix 
# using the Watanabe-Akaike Information Criterion (WAIC).
H <- 4 *diag(rep(1,p))
iter <- 100000; burnin <- floor(iter/2); thin <- 100
res <- select_k(X = X, kmax = 10, ncores = 10, H = H, nmcmc = iter, nburn = burnin, nthin = thin)
}

}
\references{
Kim, S., Lee, K., Park, S., and Lee, J. (2025+), "Eigenstructure inference for high-dimensional covariance with shrinkage inverse-Wishart prior",
	\emph{Arxiv}, URL: \url{https://arxiv.org/abs/2412.10753}.

	Watanabe, S. and Opper, M. (2010). "Asymptotic equivalence of bayes cross validation and
	widely applicable information criterion in singular learning theory", \emph{Journal of machine
	learning research}, 11(12).
	
	Gelman, A., Hwang, J., and Vehtari, A. (2014). "Understanding predictive information criteria for Bayesian models." \emph{Statistics and computing},
	24(6), 997-1016.
}
\seealso{
\code{eigenGSIW}
}
\author{
Sewon Park
}
\keyword{covariance}
\keyword{spiked}
